Grouping

+++ Create-Map.sh
@@ -1,55 +1,70 @@
+#!/bin/bash
+region = g
+proj_map = N0/12
 
-region=g
-proj_map=N0/12
+CaseNumber = 1 # Select the case number
+CurrentDirectory = !pwd
 
-# Select Model - Case1 
-# Case1Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case1/
-# Case1OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case1/
 
-# Select Model - Case2 
-#Case2Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case2/
-#Case2OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case2/
-
-# Select Model - Case3 
-# Case3Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case3/
-# Case3OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case3/
-
-# Select Model - Case4 
-Case4Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case4/
-Case4OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case4/
-
-# Only for Case 1,2,3
-# AgeGridDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Agegrid/
-
-# Only for Case 4
-AgeGridDirM21=/Users/omer/Documents/Programming/PyGplates/ForPublication/Agegrid-M21/
-
-# For Cases 1,2,3
-# CratonicShapesRootM21NNR=/Users/omer/Documents/Programming/PyGplates/ForPublication/Cratonic-Shapes-M21-NNR
-
-# Only for Case 4
-CratonicShapesRootM21=/Users/omer/Documents/Programming/PyGplates/ForPublication/Cratonic-Shapes-M21
+if [ $CaseNumber -eq 1 ]
+then 
+	# Case1Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case1/
+	CaseDir=${CurrentDirectory}/Case1
+	# Case1OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case1/
+	CaseOutputDir=${CurrentDirectory}/Output-Case1
+	AgeGridDir=${CurrentDirectory}/Agegrid-M21-NNR
+	CratonicShapes=${CurrentDirectory}/Cratonic-Shapes-M21-NNR
+	root_for_Kimb_Data=${CurrentDirectory}/Kimberlites/M21-NNR
+	
+	
+elif [$CaseNumber -eq 2 ]
+then
+	# Case1Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case1/
+	CaseDir=${CurrentDirectory}/Case2
+	# Case1OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case1/
+	CaseOutputDir=${CurrentDirectory}/Output-Case2
+	AgeGridDir=${CurrentDirectory}/Agegrid-M21-NNR
+	CratonicShapes=${CurrentDirectory}/Cratonic-Shapes-M21-NNR
+	root_for_Kimb_Data=${CurrentDirectory}/Kimberlites/M21-NNR
+	
+elif [$CaseNumber -eq 3 ]
+then
+	# Case1Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case1/
+	CaseDir=${CurrentDirectory}/Case3
+	# Case1OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case1/
+	CaseOutputDir=${CurrentDirectory}/Output-Case3
+	AgeGridDir=${CurrentDirectory}/Agegrid-M21-NNR
+	CratonicShapes=${CurrentDirectory}/Cratonic-Shapes-M21-NNR
+	root_for_Kimb_Data=${CurrentDirectory}/Kimberlites/M21-NNR
+	
+elif [$CaseNumber -eq 4 ]
+then
+	# Case1Dir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Case1/
+	CaseDir=${CurrentDirectory}/Case4
+	# Case1OutputDir=/Users/omer/Documents/Programming/PyGplates/ForPublication/Output-Case1/
+	CaseOutputDir=${CurrentDirectory}/Output-Case4
+	AgeGridDir=${CurrentDirectory}/Agegrid-M21
+	CratonicShapes=${CurrentDirectory}/Cratonic-Shapes-M21
+	root_for_Kimb_Data=${CurrentDirectory}/Kimberlites/M21
+fi
 
 #For Cases 1,2,3
 # root_for_Kimb_Data=/Users/omer/Documents/Programming/PyGplates/Kimberlites
 
-# Only for Case 4
-root_for_Kimb_Data_M21=/Users/omer/Documents/Programming/PyGplates/ForPublication/Kimberlites/M21
+for Age in `seq 0 20 200`; do	
+echo $Age
 
-for timeMa in `seq 0 20 200`; do	
-
-echo $timeMa
-
-filexyz=${Case4Dir}/Case4-Radial-Heat-Advection-Between-322km-and-CMB-Averaged-${timeMa}-Ma.xyz
+# filexyz=${Case4Dir}/Case4-Radial-Heat-Advection-Between-322km-and-CMB-Averaged-${timeMa}-Ma.xyz
+filexyz=${CaseDir}/${CaseNumber}-Radial-Heat-Advection-Between-322km-and-CMB-Averaged-${Age}-Ma.xyz
 
 echo $filexyz
 
 # Introduce names for grid files and output .ps file to be converted to PDF and/or PNG/JPG
-psfile=${Case4OutputDir}Case4-Radial-Heat-Advection-Between-322km-and-CMB-Averaged-${timeMa}-Ma.ps
-medianfile=${Case4OutputDir}Case4_Radial-Heat-Advection_between_322km_and_CMB_averaged_${timeMa}-Ma.median
-gridFile=${Case4OutputDir}Case4_Radial-Heat-Advection_between_322km_and_CMB_averaged_${timeMa}-Ma.nc
-gridFile_1_0=${Case4OutputDir}Case4_Radial-Heat-Advection_between_322km_and_CMB_averaged_1_and_0_${timeMa}-Ma.nc
-gridFileMasked=${Case4OutputDir}Case4_Radial-Heat-Advection_between_322km_and_CMB_averaged_${timeMa}masked-Ma.nc
+psfile=${CaseOutputDir}/${CaseNumber}-Radial-Heat-Advection-Between-322km-and-CMB-Averaged-${Age}-Ma.ps
+medianfile=${CaseOutputDir}/${CaseNumber}_Radial-Heat-Advection_between_322km_and_CMB_averaged_${Age}-Ma.median
+gridFile=${CaseOutputDir}/${CaseNumber}_Radial-Heat-Advection_between_322km_and_CMB_averaged_${Age}-Ma.nc
+gridFile_1_0=${CaseOutputDir}/${CaseNumber}_Radial-Heat-Advection_between_322km_and_CMB_averaged_1_and_0_${Age}-Ma.nc
+gridFileMasked=${CaseOutputDir}/${CaseNumber}_Radial-Heat-Advection_between_322km_and_CMB_averaged_${Age}masked-Ma.nc
 
 
 # If Reconstruction is M21 Merdith et al., (2021)
@@ -58,59 +73,56 @@ gridFileMasked=${Case4OutputDir}Case4_Radial-Heat-Advection_between_322km_and_CM
 
 
 # agegrid=${AgeGridDir}agegridsampled_${timeMa}-Ma.grd
-agegrid=${AgeGridDirM21}/agegrid_final_with_continents_${timeMa}.grd
+if [ $CaseNumber -eq 4 ]
+then
+	maskFileRd2=mask_Rd_2${Age}.nc
+	CratonicShapesPlot=${CratonicShapes}/Cratons_reconstructed_M21_${Age}.00Ma.xy
+	gmt grdmask -Rd -I0.1 $CratonicShapesPlot -N0/1/1 -G$maskFileRd2 
+	agegrid=${AgeGridDir}/agegrid_final_with_continents_${AgeMa}.grd
+	gmt blockmedian $filexyz -Rd -I0.1 -V > $medianfile
+	gmt surface $medianfile -I0.1 -R${region} -V -G$gridFile # Resolution is 0.1 deg.
+	gmt grdclip $gridFile -Rd -G$gridFile_1_0 -Sa0.99/1 -Sb0.99/0 -V # Set values equal and above 1 as 1, others as 0.
+	rm  $medianfile # remove temporary grid file
+	gmt grdmath $maskFileRd2 $gridFile_1_0 MUL = $gridFileMasked
+	gmt grdmath $agegrid -0.1 GT = Ocean.grd 
+	KimberlitesXY=${root_for_Kimb_Data}/${Age}Ma_kimberlite_locations-M21.xy
+	
+fi
+
 maskFile=mask_${timeMa}.nc
 maskFileRd=mask_Rd_${timeMa}.nc
-maskFileRd2=mask_Rd_2${timeMa}.nc
 
 maskFileM21=mask_M21_${timeMa}.nc	
 maskFileM21Rd=mask_M21_Rd_${timeMa}.nc
 
 
-gmt blockmedian $filexyz -Rd -I0.1 -V > $medianfile
-gmt surface $medianfile -I0.1 -R${region} -V -G$gridFile # Resolution is 0.1 deg.
 # gmt sphinterpolate $medianfile -I0.1 -R${region} -V -G$gridFile # Resolution is 0.1 deg.
 
 # gmt grdedit $gridFile -R-180/180/-90/90
-gmt grdclip $gridFile -Rd -G$gridFile_1_0 -Sa0.99/1 -Sb0.99/0 -V # Set values equal and above 1 as 1, others as 0.
 
-rm  $medianfile # remove temporary grid file
-
-CoastlinesPlot=${CratonicShapesRootM21}/Cratons_reconstructed_M21_${timeMa}.00Ma.xy
-
-gmt grdmask -Rd -I0.1 $CoastlinesPlot -N0/1/1 -G$maskFileRd2 
 
 #Below for reconstruction M21 
 #gmt grdmask -Rd -I0.1 $CratonFile -N0/1/1 -G$maskFileM21 
 #gmt grdmask -Rd -I0.1 $CratonFile -N-1/1/1 -G$maskFileM21Rd 
 
 
-gmt grdmath $maskFileRd2 $gridFile_1_0 MUL = $gridFileMasked
-
-
-gmt grdmath $agegrid -0.1 GT = Ocean.grd 
-
-CoastlinesPlot=${CratonicShapesRootM21}/Coastlines-M21-${timeMa}.00Ma.xy
+# CoastlinesPlot=${CratonicShapesRootM21}/Coastlines-M21-${timeMa}.00Ma.xy
 
 gmt grdimage -R${region} -J${proj_map} $gridFile -CColourmap-RHA.cpt -t50 -V -K > $psfile
-
-gmt psxy $CoastlinesPlot -R${region} -J${proj_map} -Gblack -O -V -K >> $psfile 
-
+gmt psxy $CratonicShapesPlot -R${region} -J${proj_map} -Gblack -O -V -K >> $psfile 
 gmt grdimage -R${region} -J${proj_map} $gridFile -CColourmap-RHA.cpt -t70 -O -V -K >> $psfile
 
-CoastlinesPlot=${CratonicShapesRootM21}/reconstructed_MerdithCratons_${timeMa}.00Ma.xy
+# CoastlinesPlot=${CratonicShapesRootM21}/reconstructed_MerdithCratons_${timeMa}.00Ma.xy
 
-gmt psxy $CoastlinesPlot -R${region} -J${proj_map}  -W0.2p,black -Ggreen -O -V -K >> $psfile 
-
+gmt psxy $CratonicShapesPlot -R${region} -J${proj_map}  -W0.2p,black -Ggreen -O -V -K >> $psfile 
 gmt grdimage -R${region} -J${proj_map} $gridFile -CColourmap-RHA.cpt -t40 -O -V -K >> $psfile
-
 gmt grdimage -R${region} -J${proj_map} Ocean.grd  -COcean.cpt -t65 -O -V -K >> $psfile
-
 gmt grdimage -R${region} -J${proj_map} $gridFileMasked -CColourmap-RHA_Masked.cpt -t60 -O -V -K >> $psfile
 
 # Add reconstructed kimberlites
 # KimberlitesXY=${root_for_Kimb_Data}/XYreconstructed_${timeMa}.00Ma.xy
-KimberlitesXY=${root_for_Kimb_Data_M21}/${timeMa}Ma_kimberlite_locations-M21.xy
+
+# KimberlitesXY=${root_for_Kimb_Data}/${Age}Ma_kimberlite_locations-M21.xy
 gmt psxy $KimberlitesXY -R${region} -J${proj_map} -Sc0.17 -W0.3p,black -Gwhite -t5 -O -V -K >> $psfile #1)
 gmt psxy $KimberlitesXY -R${region} -J${proj_map} -Sc0.08 -W0.06p,black -Gmagenta  -t3 -O -V -K >> $psfile #2)
 gmt psxy $KimberlitesXY -R${region} -J${proj_map} -Sc0.03 -W0.06p,black -Gblack -t3 -O -V -K >> $psfile #3)
